requirements.txt
pandas>=1.0.0
numpy>=1.18.0
scikit-learn>=0.22.0
tabulate>=0.8.7
matplotlib>=3.1.0
seaborn>=0.10.0
streamlit>=1.22.0
plotly>=5.14.0
sqlalchemy>=1.4.0
openai>=1.10.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0 